Create TelemetryView component. Also disable the toolbar frame button for telemetry objects.
